bot/requirements.txt
aiohttp==3.10.11
aiofiles==23.2.1
python-dotenv==1.0.0
pydantic==2.5.3
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Microsoft Teams Bot Framework
botbuilder-core==4.17.1
botbuilder-schema==4.17.1
botbuilder-integration-aiohttp==4.17.1
